requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.1.0
requests==2.32.5
gunicorn==21.2.0

